# resources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Cronograma Capilar",
    "version": "1.0.0",
    "description": "API para acompanhamento de cronograma capilar. Autenticação via JWT."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "hairType": { "type": "string" }
        }
      },
      "Cream": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "usage": { "type": "string" }
        }
      },
      "Treatment": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "Wash": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "creamId": { "type": "integer" },
          "treatment": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "result": { "type": "string" }
        }
      },
      "Progress": {
        "type": "object",
        "properties": {
          "totalWashes": { "type": "integer" },
          "byTreatment": {
            "type": "object",
            "additionalProperties": { "type": "integer" }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/user/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "hairType": { "type": "string" }
                },
                "required": ["username", "password", "hairType"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token JWT", "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" } } } } } },
          "400": { "description": "Erro de autenticação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/hair/types": {
      "get": {
        "summary": "Listar tipos de cabelo",
        "responses": {
          "200": { "description": "Tipos de cabelo", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } } }
        }
      }
    },
    "/hair": {
      "get": {
        "summary": "Consultar tipo de cabelo do usuário",
        "responses": {
          "200": { "description": "Tipo de cabelo", "content": { "application/json": { "schema": { "type": "object", "properties": { "hairType": { "type": "string" } } } } } },
          "404": { "description": "Não encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "summary": "Atualizar tipo de cabelo do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "hairType": { "type": "string" } }, "required": ["hairType"] }
            }
          }
        },
        "responses": {
          "200": { "description": "Tipo de cabelo atualizado", "content": { "application/json": { "schema": { "type": "object", "properties": { "hairType": { "type": "string" } } } } } },
          "400": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/cream": {
      "post": {
        "summary": "Cadastrar creme",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" }, "usage": { "type": "string" } }, "required": ["name", "usage"] }
            }
          }
        },
        "responses": {
          "201": { "description": "Creme cadastrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cream" } } } },
          "400": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "summary": "Listar cremes",
        "responses": {
          "200": { "description": "Lista de cremes", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cream" } } } } }
        }
      }
    },
    "/cream/{id}": {
      "get": {
        "summary": "Buscar creme por ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Creme encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cream" } } } },
          "404": { "description": "Creme não encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/cream/search/{name}": {
      "get": {
        "summary": "Buscar creme por nome",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Lista de cremes encontrados", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cream" } } } } }
        }
      }
    },
    "/treatment": {
      "get": {
        "summary": "Listar tratamentos",
        "responses": {
          "200": { "description": "Lista de tratamentos", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Treatment" } } } } }
        }
      }
    },
    "/treatment/types": {
      "get": {
        "summary": "Listar tipos de cronograma",
        "responses": {
          "200": { "description": "Tipos de cronograma", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } } }
        }
      }
    },
    "/treatment/{name}": {
      "get": {
        "summary": "Buscar tratamento por nome",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tratamento encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Treatment" } } } },
          "404": { "description": "Tratamento não encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/wash": {
      "post": {
        "summary": "Registrar lavagem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "creamId": { "type": "integer" },
                  "treatment": { "type": "string" },
                  "date": { "type": "string", "format": "date" },
                  "result": { "type": "string" }
                },
                "required": ["creamId", "treatment", "date", "result"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Lavagem registrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Wash" } } } },
          "400": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "summary": "Listar lavagens do usuário",
        "responses": {
          "200": { "description": "Lista de lavagens", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Wash" } } } } }
        }
      }
    },
    "/wash/last": {
      "get": {
        "summary": "Buscar última lavagem",
        "responses": {
          "200": { "description": "Última lavagem", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Wash" } } } },
          "404": { "description": "Nenhuma lavagem encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/progress": {
      "get": {
        "summary": "Consultar progresso do tratamento",
        "responses": {
          "200": { "description": "Progresso do tratamento", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Progress" } } } }
        }
      }
    }
  }
}